Update AI_Interface.py
Primary Update - Moved existing Settings to a dedicated Settings Menu + Created new Settings. Current available settings:
Audio Settings:
- Input Device
- Output Device
- Auto-Voice Enabling > Turns mic back on after AI response (Prone to crashes at this time which are being reviewed.)
AI Settings:
- AI reader - Choose Edge reader.
- AI Model - You can now change the ollama model
- Restore Base Context - For Backup Purposes.
Memory Specific Settings:
- Session Only Memory - If you don't want the conversation to be saved (Limited to up to 10 messages)
- Memory Limit - You can decide how many messages the AI remembers. While theoretically this would be infinite, due to token limitations of LLM's, there will always be a limit to the generations. Please note as well that the more memory the AI has, the higher chances of hallucinations occuring is.
- Clear All Saved Memories - This clears memories saved on the system and clears the session based memory upon usage so the conversation will start anew.
Visual Settings:
- Boot Up in Dark Mode - To be able to decide if you get flashbanged or not at night is nice!


Additional Update -Updated the layout and function of the buttons of the home screen. Current function of the bottom buttons would now be the following:

Clear Chat Log - To clear the chat log (This is only visual and does not clear memory)
Context Manager - A direct way to edit the context.
Settings - To enter the settings menu
Addons Menu - Currently is useless (just opens a menu saying that it is useless), but meant for accessing future additional features on the program. Current Ideas are Vtube Integration, possible twitch chat reader and integration with other programs or features, noting that this is not a guarantee to be added. Just an idea that I had.